Keep dist-root in @INC beyond Perl 5.26
The only thing that makes sense is to ensure CWD of some description is
in @INC at some stage before plugins are actually loaded, because
that's the traditional expectation.

Basically, this option is the only option that works. There's no viable
way to scope @INC changes to anything narrower without breaking the
usage entirely.

There's no sane way to limit the depth of @INC changes either, even if
we wanted to.

And if you have vulnerable code in your project source tree that you're
worried about, you're pretty much beyond screwed already, and you're
about to be screwing CPAN when you ship next, and no amount of @INC
magic will save you.

Bug: https://bugs.gentoo.org/613586
Bug: https://github.com/rjbs/Dist-Zilla/issues/581
Bug: https://github.com/rjbs/Dist-Zilla/pull/590
Bug: https://github.com/rjbs/Dist-Zilla/pull/600